style: domain_for_infra -> domain_for_user_input

+++ src/form/domain/form_id.rs
@@ -2,5 +2,5 @@ use typed_builder::TypedBuilder;
 
 #[derive(TypedBuilder)]
 pub struct FormId {
-    form_id: u8,
+    form_id: i32,
 }
